Refactoring distribution/distribution.py
This file was becoming too large and unclear what it housed. Split it
into pieces.
