+++ llama_stack/distribution/distribution.py
@@ -5,30 +5,11 @@
 # the root directory of this source tree.
 
 import importlib
-import inspect
 from typing import Dict, List
 
 from pydantic import BaseModel
 
-from llama_stack.apis.agents import Agents
-from llama_stack.apis.inference import Inference
-from llama_stack.apis.memory import Memory
-from llama_stack.apis.memory_banks import MemoryBanks
-from llama_stack.apis.models import Models
-from llama_stack.apis.safety import Safety
-from llama_stack.apis.shields import Shields
-from llama_stack.apis.telemetry import Telemetry
-
-from .datatypes import Api, ApiEndpoint, ProviderSpec, remote_provider_spec
-
-# These are the dependencies needed by the distribution server.
-# `llama-stack` is automatically installed by the installation script.
-SERVER_DEPENDENCIES = [
-    "fastapi",
-    "fire",
-    "httpx",
-    "uvicorn",
-]
+from llama_stack.providers.datatypes import Api, ProviderSpec, remote_provider_spec
 
 
 def stack_apis() -> List[Api]:
@@ -57,45 +38,7 @@ def builtin_automatically_routed_apis() -> List[AutoRoutedApiInfo]:
     ]
 
 
-def api_endpoints() -> Dict[Api, List[ApiEndpoint]]:
-    apis = {}
-
-    protocols = {
-        Api.inference: Inference,
-        Api.safety: Safety,
-        Api.agents: Agents,
-        Api.memory: Memory,
-        Api.telemetry: Telemetry,
-        Api.models: Models,
-        Api.shields: Shields,
-        Api.memory_banks: MemoryBanks,
-    }
-
-    for api, protocol in protocols.items():
-        endpoints = []
-        protocol_methods = inspect.getmembers(protocol, predicate=inspect.isfunction)
-
-        for name, method in protocol_methods:
-            if not hasattr(method, "__webmethod__"):
-                continue
-
-            webmethod = method.__webmethod__
-            route = webmethod.route
-
-            if webmethod.method == "GET":
-                method = "get"
-            elif webmethod.method == "DELETE":
-                method = "delete"
-            else:
-                method = "post"
-            endpoints.append(ApiEndpoint(route=route, method=method, name=name))
-
-        apis[api] = endpoints
-
-    return apis
-
-
-def api_providers() -> Dict[Api, Dict[str, ProviderSpec]]:
+def get_provider_registry() -> Dict[Api, Dict[str, ProviderSpec]]:
     ret = {}
     routing_table_apis = set(
         x.routing_table_api for x in builtin_automatically_routed_apis()
